invented h2stdinc.h which acts as the main header for system headers, for our basic types, and does compile time checks for certain type sizes. all files must include h2stdinc.h first, h2def.h is the second from now on. besides, we now depend on stdint.h and/or inttypes.h, because some structures require int32_t (they used to require long) and this should help with 64 bit compatibility.
git-svn-id: svn://svn.code.sf.net/p/hhexen/code/trunk/hhexen@72 989aa981-fe4f-0410-aa6a-c19d413bcdcd

+++ base/i_sound.c
@@ -1,7 +1,7 @@
-// $Id: i_sound.c,v 1.4 2008-06-17 09:20:10 sezero Exp $
+// $Id: i_sound.c,v 1.5 2008-06-17 13:40:30 sezero Exp $
 
 
-#include <stdio.h>
+#include "h2stdinc.h"
 #include <math.h>       // pow()
 #include <pthread.h>
 #include "h2def.h"
@@ -122,7 +122,7 @@ typedef struct
 {
     short a;            // always 3
     short freq;         // always 11025
-    long length;        // sample length
+    int32_t length;     // sample length
     unsigned char firstSample;
 } Sample;
